+++ app/Http/Requests/UpdateUserAccount.php
@@ -25,7 +25,7 @@ class UpdateUserAccount extends FormRequest
     public function rules()
     {
         return [
-            'email'             => 'sometimes|required|max:255|email|unique:users',
+            'email'             => 'sometimes|required|max:255|email',
             'current_password'  => 'sometimes|required',
             'new_password'      => 'required_with:current_password|min:8|confirmed',
         ];